+++ src/compas_rhino/scene/meshobject.py
@@ -8,13 +8,11 @@ import scriptcontext as sc  # type: ignore
 import compas_rhino.objects
 from compas.colors import Color
 from compas.geometry import Line
-from compas.geometry import Point
 from compas.geometry import centroid_points
 from compas.scene import MeshObject
 from compas_rhino.conversions import line_to_rhino
 from compas_rhino.conversions import point_to_rhino
-
-# from compas_rhino.conversions import transformation_to_rhino
+from compas_rhino.conversions import transformation_to_rhino
 from compas_rhino.conversions import vertices_and_faces_to_rhino
 
 from .helpers import ngon
@@ -56,16 +54,9 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         super(RhinoMeshObject, self).__init__(**kwargs)
         self.disjoint = disjoint
         self._guid_mesh = None
-        self._guids_faces = None
-        self._guids_edges = None
-        self._guids_vertices = None
-        self._guids_vertexnormals = None
-        self._guids_facenormals = None
-        self._guids_vertexlabels = None
-        self._guids_edgelabels = None
-        self._guids_facelabels = None
-        self._guids_spheres = None
-        self._guids_pipes = None
+        self._guid_face = {}
+        self._guid_edge = {}
+        self._guid_vertex = {}
         self.vertexgroup = vertexgroup
         self.edgegroup = edgegroup
         self.facegroup = facegroup
@@ -92,7 +83,7 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         None
 
         """
-        compas_rhino.objects.delete_objects(self._guids_vertices, purge=True)
+        compas_rhino.objects.delete_objects(self._guid_vertex, purge=True)
 
     def clear_edges(self):
         """Delete all edges drawn by this scene object.
@@ -102,7 +93,7 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         None
 
         """
-        compas_rhino.objects.delete_objects(self._guids_edges, purge=True)
+        compas_rhino.objects.delete_objects(self._guid_edge, purge=True)
 
     def clear_faces(self):
         """Delete all faces drawn by this scene object.
@@ -112,57 +103,7 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         None
 
         """
-        compas_rhino.objects.delete_objects(self._guids_faces, purge=True)
-
-    def clear_vertexnormals(self):
-        """Delete all vertex normals drawn by this scene object.
-
-        Returns
-        -------
-        None
-
-        """
-        compas_rhino.objects.delete_objects(self._guids_vertexnormals, purge=True)
-
-    def clear_facenormals(self):
-        """Delete all face normals drawn by this scene object.
-
-        Returns
-        -------
-        None
-
-        """
-        compas_rhino.objects.delete_objects(self._guids_facenormals, purge=True)
-
-    def clear_vertexlabels(self):
-        """Delete all vertex labels drawn by this scene object.
-
-        Returns
-        -------
-        None
-
-        """
-        compas_rhino.objects.delete_objects(self._guids_vertexlabels, purge=True)
-
-    def clear_edgelabels(self):
-        """Delete all edge labels drawn by this scene object.
-
-        Returns
-        -------
-        None
-
-        """
-        compas_rhino.objects.delete_objects(self._guids_edgelabels, purge=True)
-
-    def clear_facelabels(self):
-        """Delete all face labels drawn by this scene object.
-
-        Returns
-        -------
-        None
-
-        """
-        compas_rhino.objects.delete_objects(self._guids_facelabels, purge=True)
+        compas_rhino.objects.delete_objects(self._guid_face, purge=True)
 
     # ==========================================================================
     # draw
@@ -193,17 +134,16 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             attr = self.compile_attributes()
 
             vertexcolors = []
-            if len(self.vertexcolor):
-                vertexcolors = [self.vertexcolor[vertex] for vertex in self.mesh.vertices()]
+            if len(self.vertexcolor):  # type: ignore
+                vertexcolors = [self.vertexcolor[vertex] for vertex in self.mesh.vertices()]  # type: ignore
 
             facecolors = []
-            if len(self.facecolor):
-                facecolors = [self.facecolor[face] for face in self.mesh.faces()]
+            if len(self.facecolor):  # type: ignore
+                facecolors = [self.facecolor[face] for face in self.mesh.faces()]  # type: ignore
 
             vertex_index = self.mesh.vertex_index()
-            vertex_xyz = self.vertex_xyz
 
-            vertices = [vertex_xyz[vertex] for vertex in self.mesh.vertices()]
+            vertices = [self.mesh.vertex_attributes(vertex, "xyz") for vertex in self.mesh.vertices()]
             faces = [[vertex_index[vertex] for vertex in self.mesh.face_vertices(face)] for face in self.mesh.faces()]
 
             geometry = vertices_and_faces_to_rhino(
@@ -215,7 +155,7 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
                 disjoint=self.disjoint,
             )
 
-            # geometry.Transform(transformation_to_rhino(self.worldtransformation))
+            geometry.Transform(transformation_to_rhino(self.worldtransformation))
 
             self._guid_mesh = sc.doc.Objects.AddMesh(geometry, attr)
             if self.group:
@@ -224,13 +164,10 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             self._guids.append(self._guid_mesh)
 
         elif self.show_faces:
-            self._guids += self.draw_faces()
+            self.draw_faces()
 
-        if self.show_vertices:
-            self._guids += self.draw_vertices()
-
-        if self.show_edges:
-            self._guids += self.draw_edges()
+        self.draw_vertices()
+        self.draw_edges()
 
         return self.guids
 
@@ -247,15 +184,18 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
 
         vertices = list(self.mesh.vertices()) if self.show_vertices is True else self.show_vertices or []
 
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         if vertices:
             for vertex in vertices:
                 name = "{}.vertex.{}".format(self.mesh.name, vertex)
-                color = self.vertexcolor[vertex]
+                color = self.vertexcolor[vertex]  # type: ignore
                 attr = self.compile_attributes(name=name, color=color)
 
-                point = point_to_rhino(self.vertex_xyz[vertex])
+                geometry = point_to_rhino(self.mesh.vertex_attributes(vertex, "xyz"))
+                geometry.Transform(transformation)
 
-                guid = sc.doc.Objects.AddPoint(point, attr)
+                guid = sc.doc.Objects.AddPoint(geometry, attr)
                 guids.append(guid)
 
         if guids:
@@ -264,7 +204,9 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             elif self.group:
                 self.add_to_group(self.group, guids)
 
-        self._guids_vertices = guids
+        self._guid_vertex = dict(zip(guids, vertices))
+
+        self._guids += guids
         return guids
 
     def draw_edges(self):
@@ -280,15 +222,20 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
 
         edges = list(self.mesh.edges()) if self.show_edges is True else self.show_edges or []
 
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         if edges:
             for edge in edges:
                 name = "{}.edge.{}-{}".format(self.mesh.name, *edge)
-                color = self.edgecolor[edge]
+                color = self.edgecolor[edge]  # type: ignore
                 attr = self.compile_attributes(name=name, color=color)
 
-                line = Line(self.vertex_xyz[edge[0]], self.vertex_xyz[edge[1]])
+                line = self.mesh.edge_line(edge)
 
-                guid = sc.doc.Objects.AddLine(line_to_rhino(line), attr)
+                geometry = line_to_rhino(line)
+                geometry.Transform(transformation)
+
+                guid = sc.doc.Objects.AddLine(geometry, attr)
                 guids.append(guid)
 
         if guids:
@@ -297,7 +244,9 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             elif self.group:
                 self.add_to_group(self.group, guids)
 
-        self._guids_edges = guids
+        self._guid_edge = dict(zip(guids, edges))
+
+        self._guids += guids
         return guids
 
     def draw_faces(self):
@@ -313,17 +262,22 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
 
         faces = list(self.mesh.faces()) if self.show_faces is True else self.show_faces or []
 
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         if faces:
             for face in faces:
                 name = "{}.face.{}".format(self.mesh.name, face)
-                color = self.facecolor[face]
+                color = self.facecolor[face]  # type: ignore
                 attr = self.compile_attributes(name=name, color=color)
 
-                vertices = [self.vertex_xyz[vertex] for vertex in self.mesh.face_vertices(face)]  # type: ignore
+                vertices = [self.mesh.vertex_attributes(vertex, "xyz") for vertex in self.mesh.face_vertices(face)]  # type: ignore
                 facet = ngon(len(vertices))
 
                 if facet:
-                    guid = sc.doc.Objects.AddMesh(vertices_and_faces_to_rhino(vertices, [facet]), attr)
+                    geometry = vertices_and_faces_to_rhino(vertices, [facet])
+                    geometry.Transform(transformation)
+
+                    guid = sc.doc.Objects.AddMesh(geometry, attr)
                     guids.append(guid)
 
         if guids:
@@ -332,7 +286,9 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             elif self.group:
                 self.add_to_group(self.group, guids)
 
-        self._guids_faces = guids
+        self._guid_face = dict(zip(guids, faces))
+
+        self._guids += guids
         return guids
 
     # ==========================================================================
@@ -365,14 +321,17 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
 
         self.vertexcolor = color
 
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         for vertex in text:
             name = "{}.vertex.{}.label".format(self.mesh.name, vertex)  # type: ignore
-            color = self.vertexcolor[vertex]
+            color = self.vertexcolor[vertex]  # type: ignore
             attr = self.compile_attributes(name=name, color=color)
 
-            point = point_to_rhino(self.vertex_xyz[vertex])
+            location = point_to_rhino(self.mesh.vertex_attributes(vertex, "xyz"))
+            location.Transform(transformation)
 
-            dot = Rhino.Geometry.TextDot(str(text[vertex]), point)  # type: ignore
+            dot = Rhino.Geometry.TextDot(str(text[vertex]), location)  # type: ignore
             dot.FontHeight = fontheight
             dot.FontFace = fontface
 
@@ -383,7 +342,7 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             self.add_to_group(group, guids)
 
         self._guids_vertexlabels = guids
-
+        self._guids += guids
         return guids
 
     def draw_edgelabels(self, text, color=None, group=None, fontheight=10, fontface="Arial Regular"):
@@ -412,15 +371,18 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
 
         self.edgecolor = color
 
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         for edge in text:
             name = "{}.edge.{}-{}".format(self.mesh.name, *edge)  # type: ignore
-            color = self.edgecolor[edge]
+            color = self.edgecolor[edge]  # type: ignore
             attr = self.compile_attributes(name="{}.label".format(name), color=color)
 
-            line = Line(self.vertex_xyz[edge[0]], self.vertex_xyz[edge[1]])
-            point = point_to_rhino(line.midpoint)
+            line = self.mesh.edge_line(edge)
+            location = point_to_rhino(line.midpoint)
+            location.Transform(transformation)
 
-            dot = Rhino.Geometry.TextDot(str(text[edge]), point)  # type: ignore
+            dot = Rhino.Geometry.TextDot(str(text[edge]), location)  # type: ignore
             dot.FontHeight = fontheight
             dot.FontFace = fontface
 
@@ -431,7 +393,7 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             self.add_to_group(group, guids)
 
         self._guids_edgelabels = guids
-
+        self._guids += guids
         return guids
 
     def draw_facelabels(self, text, color=None, group=None, fontheight=10, fontface="Arial Regular"):
@@ -458,15 +420,18 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         """
         guids = []
 
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         for face in text:
             name = "{}.face.{}.label".format(self.mesh.name, face)  # type: ignore
-            color = self.facecolor[face]
+            color = self.facecolor[face]  # type: ignore
             attr = self.compile_attributes(name=name, color=color)
 
-            points = [self.vertex_xyz[vertex] for vertex in self.mesh.face_vertices(face)]  # type: ignore
-            point = point_to_rhino(centroid_points(points))
+            points = [self.mesh.vertex_attributes(vertex, "xyz") for vertex in self.mesh.face_vertices(face)]  # type: ignore
+            location = point_to_rhino(centroid_points(points))
+            location.Transform(transformation)
 
-            dot = Rhino.Geometry.TextDot(str(text[face]), point)  # type: ignore
+            dot = Rhino.Geometry.TextDot(str(text[face]), location)  # type: ignore
             dot.FontHeight = fontheight
             dot.FontFace = fontface
 
@@ -477,23 +442,21 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
             self.add_to_group(group, guids)
 
         self._guids_facelabels = guids
-
+        self._guids += guids
         return guids
 
     # ==========================================================================
     # draw normals
     # ==========================================================================
 
-    def draw_vertexnormals(self, vertices=None, color=(0, 255, 0), scale=1.0, group=None):
+    def draw_vertexnormals(self, color=None, scale=1.0, group=None):
         """Draw the normals at the vertices of the mesh.
 
         Parameters
         ----------
-        vertices : list[int], optional
-            A selection of vertex normals to draw.
-            Default is to draw all vertex normals.
         color : tuple[int, int, int] | tuple[float, float, float] | :class:`compas.colors.Color`, optional
             The color specification of the normal vectors.
+            If no  color is specified, the color of the corresponding vertex is used.
         scale : float, optional
             Scale factor for the vertex normals.
         group : str, optional
@@ -507,35 +470,40 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         """
         guids = []
 
+        vertices = list(self.mesh.vertices()) if self.show_vertices is True else self.show_vertices or []
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         color = Color.coerce(color)
 
-        for vertex in vertices or self.mesh.vertices():  # type: ignore
+        for vertex in vertices:
             name = "{}.vertex.{}.normal".format(self.mesh.name, vertex)  # type: ignore
-            attr = self.compile_attributes(name=name, color=color)
+            attr = self.compile_attributes(name=name, color=color or self.vertexcolor[vertex])  # type: ignore
 
-            point = Point(*self.vertex_xyz[vertex])
+            point = self.mesh.vertex_point(vertex)
             normal = self.mesh.vertex_normal(vertex)  # type: ignore
+            line = Line.from_point_and_vector(point, normal * scale)
 
-            guid = sc.doc.Objects.AddLine(point_to_rhino(point), point_to_rhino(point + normal * scale), attr)
+            geometry = line_to_rhino(line)
+            geometry.Transform(transformation)
+
+            guid = sc.doc.Objects.AddLine(geometry, attr)
             guids.append(guid)
 
         if group:
             self.add_to_group(group, guids)
 
         self._guids_vertexnormals = guids
-
+        self._guids += guids
         return guids
 
-    def draw_facenormals(self, faces=None, color=(0, 255, 255), scale=1.0, group=None):
+    def draw_facenormals(self, color=None, scale=1.0, group=None):
         """Draw the normals of the faces.
 
         Parameters
         ----------
-        faces : list[int], optional
-            A selection of face normals to draw.
-            Default is to draw all face normals.
         color : tuple[int, int, int] | tuple[float, float, float] | :class:`compas.colors.Color`, optional
             The color specification of the normal vectors.
+            If no color is specified, the color of the corresponding face is used.
         scale : float, optional
             Scale factor for the face normals.
         group : str, optional
@@ -549,16 +517,23 @@ class RhinoMeshObject(RhinoSceneObject, MeshObject):
         """
         guids = []
 
+        faces = list(self.mesh.faces()) if self.show_faces is True else self.show_faces or []
+        transformation = transformation_to_rhino(self.worldtransformation)
+
         color = Color.coerce(color)
 
-        for face in faces or self.mesh.faces():  # type: ignore
+        for face in faces:
             name = "{}.face.{}.normal".format(self.mesh.name, face)  # type: ignore
-            attr = self.compile_attributes(name=name, color=color)
+            attr = self.compile_attributes(name=name, color=color or self.facecolor[face])  # type: ignore
 
-            point = Point(*centroid_points([self.vertex_xyz[vertex] for vertex in self.mesh.face_vertices(face)]))  # type: ignore
-            normal = self.mesh.face_normal(face)  # type: ignore
+            point = self.mesh.face_centroid(face)
+            normal = self.mesh.face_normal(face)
+            line = Line.from_point_and_vector(point, normal * scale)
 
-            guid = sc.doc.Objects.AddLine(point_to_rhino(point), point_to_rhino(point + normal * scale), attr)
+            geometry = line_to_rhino(line)
+            geometry.Transform(transformation)
+
+            guid = sc.doc.Objects.AddLine(geometry, attr)
             guids.append(guid)
 
         if group: